decompose exception classes
