+++ src/Publiux/laravelcdn/Exceptions/CdnException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Publiux\laravelcdn\Exceptions;
+
+/**
+ * @author Mahmoud Zalt <mahmoud@vinelab.com>
+ */
+class CdnException extends \RuntimeException
+{
+}
+
