+++ src/Publiux/laravelcdn/Exceptions/EmptyPathException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Publiux\laravelcdn\Exceptions;
+
+class EmptyPathException extends CdnException
+{
+}
