--- src/Publiux/laravelcdn/Exceptions/Exceptions.php
@@ -1,26 +1,0 @@
-<?php
-
-namespace Publiux\laravelcdn\Exceptions;
-
-/**
- * @author Mahmoud Zalt <mahmoud@vinelab.com>
- */
-class CdnException extends \RuntimeException
-{
-}
-
-class MissingConfigurationFileException extends CdnException
-{
-}
-
-class MissingConfigurationException extends CdnException
-{
-}
-
-class UnsupportedProviderException extends CdnException
-{
-}
-
-class EmptyPathException extends CdnException
-{
-}
