+++ src/Publiux/laravelcdn/Exceptions/MissingConfigurationException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Publiux\laravelcdn\Exceptions;
+
+class MissingConfigurationException extends CdnException
+{
+}
