+++ src/Publiux/laravelcdn/Exceptions/MissingConfigurationFileException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Publiux\laravelcdn\Exceptions;
+
+class MissingConfigurationFileException extends CdnException
+{
+}
