+++ src/Publiux/laravelcdn/Exceptions/UnsupportedProviderException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Publiux\laravelcdn\Exceptions;
+
+class UnsupportedProviderException extends CdnException
+{
+}
